modularized sampling functionalities
