+++ src/qrisp/jasp/interpreter_tools/interpreters/__init__.py
@@ -20,3 +20,4 @@ from qrisp.jasp.interpreter_tools.interpreters.pjit_flattening import *
 from qrisp.jasp.interpreter_tools.interpreters.environment_flattening import *
 from qrisp.jasp.interpreter_tools.interpreters.control_flow_interpretation import *
 from qrisp.jasp.interpreter_tools.interpreters.jaspr_to_gate_interpreter import *
+from qrisp.jasp.interpreter_tools.interpreters.terminal_sampling_interpreter import *
